% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_sens.R
\name{med.sens}
\alias{med.sens}
\title{Run sensitivity analysis at specified sensitivity parameter values}
\usage{
med.sens(
  effect = "IIE",
  cf.y = c(0.26, 0.26),
  cf.m = c(0.26, 0.26),
  rho = c(1, 1),
  short_list
)
}
\arguments{
\item{effect}{A \code{string} specifying the name of the effect. Options include  "IIE" (interventional indirect effect) or "IDE" (interventional direct effect)}

\item{cf.y}{(optional) A \code{numeric} vector of the R-squared sensitivity parameters for the regressions, that is, R^2_\{Y-theta_\{2,s\} ~ theta_2 -theta_\{2,s\} \} and R^2_\{bM-theta_\{1,s\} ~ theta_\{1,u,s\} -theta_\{1,s\} \}.}

\item{cf.m}{(optional) A \code{numeric} vector of the R-squared sensitivity parameters for the weights, that is, 1-R^2_\{alpha_2 ~ alpha_\{2,s\}\} and 1-R^2_\{alpha_1 ~ alpha_\{1,s\}\}.}

\item{rho}{(optional) A \code{numeric} vector of the correlation sensitivity parameters, that is, rho_\{1\} and rho_\{2\}.}

\item{short_list}{An object named "short_list" returned by the function \code{run.sensmed}.}
}
\description{
Compute estimates and confidence intervals for the bounds of the interventional indirect or direct effect at specified sensitivity parameter values.
}
