% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{run.benchmark}
\alias{run.benchmark}
\title{Obtain sensitivity parameter values based on observed covariates}
\usage{
run.benchmark(short_list, benchmark_covariates)
}
\arguments{
\item{short_list}{An object named "short_list" returned by the function \code{run.sensmed}.}

\item{benchmark_covariates}{A \code{character} vector containing the column names of the observed covariates used as the benchmark covariate.}
}
\description{
Compute gain metrics for an observed covariate (or a set of observed covariates) used as a benchmark.
}
