% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sensmed.R
\name{run.sensmed}
\alias{run.sensmed}
\title{Sensitivity analysis for causal mediation analysis}
\usage{
run.sensmed(
  data,
  treatment,
  outcome,
  mediators,
  covariates,
  learners = c("glm"),
  cf.y = c(0.26, 0.26),
  cf.m = c(0.26, 0.26),
  rho = c(1, 1),
  RV.options = list(theta.null = 0, cf = seq(0, 0.99, by = 0.01), rho = c(1, 1)),
  control = estimate_control()
)
}
\arguments{
\item{data}{A \code{data.frame} containing all analysis variables.}

\item{treatment}{The column name of the treatment variable. Currently supports binary treatment.}

\item{outcome}{The column name of the outcome variable.}

\item{mediators}{A character vector containing the column names of the mediator variables.}

\item{covariates}{A character vector containing the column names of pretreatment covariates.}

\item{learners}{A character vector of \code{mlr3superlearner} methods for estimating models. Default is \code{learners = c("glm")}. All available methods could be found at \link{https://mlr-org.com/learners.html}; examples include "ranger" (random forest), "nnet" (neural network), "lightgbm" (light implementation of gradient boosting), "xgboost" (extreme gradient boosting), "rpart" (regression tree), "glmnet" (GLM with elastic net regularization), "earth" (multivariate adaptive splines).}

\item{cf.y}{(optional) A \code{numeric} vector of the R-squared sensitivity parameters for the regressions, that is, R^2_\{Y-theta_\{2,s\} ~ theta_2 -theta_\{2,s\} \} and R^2_\{bM-theta_\{1,s\} ~ theta_\{1,u,s\} -theta_\{1,s\} \}.}

\item{cf.m}{(optional) A \code{numeric} vector of the R-squared sensitivity parameters for the weights, that is, 1-R^2_\{alpha_2 ~ alpha_\{2,s\}\} and 1-R^2_\{alpha_1 ~ alpha_\{1,s\}\}.}

\item{rho}{(optional) A \code{numeric} vector of the correlation sensitivity parameters, that is, rho_\{1\} and rho_\{2\}.}

\item{RV.options}{(optional) A list of the options for computing the robustness value specified following the template \code{RV.options = list(theta.null = 0, cf = seq(0, 0.99, by = 0.01), rho = c(1, 1))}, where
"theta.null" is the null value,
"cf" is the vector of R-squared sensitivity parameter values to search over, and
"rho" is a vector of the correlation sensitivity parameters rho_\{1\} and rho_\{2\}.}

\item{control}{(optional) A list of control parameters specified via \code{estimate_control()}. The default setting is based on R package \code{crumble}.}
}
\value{
A list containing the following:
\item{short_list}{A list containing estimates for short components.}
\item{tab_sensitivity_reporting}{A data.frame containing robustness values, and estimates, standard errors, and confidence intervals for the short parameter.}
\item{sens_res}{A data.frame containing sensitivity results at the specified sensitivity parameter values: estimates, standard errors, and confidence intervals for the short parameter (theta_s), lower bound (theta_m), and upper bound (theta_p).}
}
\description{
Implements a sensitivity analysis method for unmeasured pretreatment confounders in assessing causal mediation effects based on the debiased machine learning approach.
}
\examples{
\donttest{
    library(glue)
    library(purrr)
    library(tidyverse)
    library(origami)
    library(mlr3superlearner)
    library(mlr3extralearners)


    data("simulated_data")
    dat <- aa

    # RV and sensitivity results at specified sensitivity parameter values

    sens_M1 <- SensMed::run.sensmed(
        data = dat,
        treatment = "A",
        outcome = "Y",
        mediators = "M",
        covariates = c("X1obs", "X2obs"),
        learners = c("mean", "glm", "ranger"),
        cf.y = c(0.26, 0.26),
        cf.m = c(0.26, 0.26),
        rho = c(1, 1),
        RV.options = list(theta.null = 0, cf = seq(0, 0.99, by = 0.01), rho = c(1, 1))
    )

    # Robustness values, and estimates, standard errors, and confidence intervals for the short parameter
    sens_M1$tab_sensitivity_reporting
    # Sensitivity results at the specified sensitivity parameter values: estimates, standard errors, and confidence intervals for the short parameter (theta_s), lower bound (theta_m), and upper bound (theta_p)
    sens_M1$sens_res

}
}
