% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_contour.R
\name{run.contour}
\alias{run.contour}
\title{Make a contour plot using sensitivity parameter values}
\usage{
run.contour(
  effect = "IIE",
  target = "theta_m.90\%CI_low",
  short_list,
  bench.cf.y = NULL,
  bench.cf.m = NULL,
  rho = c(1, 1),
  cf_grid = NULL,
  y_axis = NULL,
  x_axis = NULL,
  contour_levels = NULL,
  plot_zero_confounding = FALSE,
  only_YM_confounded = FALSE
)
}
\arguments{
\item{effect}{A \code{string} specifying the name of the effect. Options include  "IIE" (interventional indirect effect) or "IDE" (interventional direct effect).}

\item{short_list}{An object named "short_list" returned by the function \code{run.sensmed}.}

\item{bench.cf.y}{(optional) A \code{numeric(1)} value to be used for both the R-squared sensitivity parameters for the regressions, to be added as an additional point on the plot.}

\item{bench.cf.m}{(optional) A \code{numeric(1)} value to be used for both the R-squared sensitivity parameters for the weights, to be added as an additional point on the plot.}

\item{rho}{(optional) A \code{numeric} vector of the correlation sensitivity parameters, that is, rho_\{1\} and rho_\{2\}.}
}
\description{
Compute estimates and confidence intervals for the bounds of the interventional indirect or direct effect at specified sensitivity parameter values.
}
